refactor; cleanup headers; diff results -> uint;

+++ gpu_utils.h
@@ -1,9 +1,6 @@
 #ifndef GPU_UTILS_H
 #define GPU_UTILS_H
 
-#include <stdio.h>
-
-
 typedef enum TaskStatus {
     TASK_DONE,
     TASK_FAILED
@@ -15,15 +12,15 @@ typedef enum TaskStatus {
 
 extern "C" unsigned int get_max_tiles_count_per_stream();
 
-extern "C" TaskStatus compare_one_image_with_others_streams(const unsigned char * const raw_left_image,
-                                                 const unsigned char * const raw_right_images,
+extern "C" TaskStatus compare_one_image_with_others_streams(const unsigned char* const raw_left_image,
+                                                 const unsigned char* const raw_right_images,
                                                  const unsigned int right_images_count,
-                                                 unsigned short int* const diff_results);
+                                                 unsigned int* const diff_results);
 
-extern "C" TaskStatus compare_one_image_with_others(const unsigned char * const raw_left_image,
-                                                      const unsigned char * const raw_right_images,
+extern "C" TaskStatus compare_one_image_with_others(const unsigned char* const raw_left_image,
+                                                      const unsigned char* const raw_right_images,
                                                       const unsigned int right_images_count,
-                                                      unsigned short *const diff_results);
+                                                      unsigned int* const diff_results);
 
 extern "C" void* gpu_backend_memory_allocator(size_t bytes);
 extern "C" void gpu_backend_memory_deallocator(void* ptr);
@@ -32,15 +29,15 @@ extern "C" void gpu_backend_memory_deallocator(void* ptr);
 
 unsigned int get_max_tiles_count_per_stream();
 
-TaskStatus compare_one_image_with_others_streams(const unsigned char * const raw_left_image,
-                                                 const unsigned char * const raw_right_images,
+TaskStatus compare_one_image_with_others_streams(const unsigned char* const raw_left_image,
+                                                 const unsigned char* const raw_right_images,
                                                  const unsigned int right_images_count,
-                                                 unsigned short int* const diff_results);
+                                                 unsigned int* const diff_results);
 
-TaskStatus compare_one_image_with_others(const unsigned char * const raw_left_image,
-                                                      const unsigned char * const raw_right_images,
+TaskStatus compare_one_image_with_others(const unsigned char* const raw_left_image,
+                                                      const unsigned char* const raw_right_images,
                                                       const unsigned int right_images_count,
-                                                      unsigned short *const diff_results);
+                                                      unsigned int* const diff_results);
 
 void* gpu_backend_memory_allocator(size_t bytes);
 void gpu_backend_memory_deallocator(void* ptr);
